Date sorting
Vorbis and FLAC date tags are informally standardized to ISO 8601,
i.e. 1999-08-16. cmus currently fails to parse this format, and won't
sort albums with this date format correctly in the tree view.

I wrote a simple date parser that can handle this format as well as
the year-only dates that MP3s use. I can't think of any other date
formats that are in widespread use for music tagging, but if anyone
can suggest any it would be easy to add support for them.