waveout: Fix buffer space calculation
There is a bug in buffer space calculation that sometimes causes
waveout_buffer_space() to return the wrong amount of free buffer space.
Eventually this triggers a BUG_ON() in buffer.c.

PCM data is not frame aligned on each buffer. This can lead to garbled
sound when playing 24-bit PCM, etc.. depending on the buffer size.