mad: Fix duration calculation
mad_header_decode() doesn't necessarily detect corrupted frame so we
have to call mad_frame_decode() instead which also decodes the header
for you.  Calling mad_header_decode() was meant to be a optimization but
since VBR files have a XING tag and it is easy to approximate duration
from CBR files this optimization was never really used.

Without this patch if there were N corrupted frames at beginning of the
file then the duration was miscalculated (N + 1) times the actual
duration.