Remove fuzzy artist name fixing, add fuzzy_artist_sort and pretty_artist_name options
Artist name fixing doesn't seem to be very useful in practice so remove
it.  Also remove raw_name from struct artist because we can always store
real artist name to artist.name and skip "the " when needed.

fuzzy_artist_sort when enabled makes sorting ignore "the " in front of
artist names.  So artists starting with "the " are not clumped together.
Defaults to false (old behavior).

pretty_artist_name when turned on changes displayed artist name so that
leading "the " is removed and ", The" is added at end the name.  This
is useful with fuzzy_artist_sort.  Defaults to false.