Support metadata in Vorbis streams
This adds support for displaying metadata in Vorbis streams. The decoder does
most of the work, so the only necessary change is notifying the player at the
start of each song. Since it is now possible for a stream track to have full
comments (as opposed to a simple SHOUTcast title string), a minor update to
the UI code determining the displayed format is also required.

This also includes support for status_display_program.

Acked-by: Frank Terbeck <ft@bewatermyfriend.org>