Fix buffer length calculation in editable.c
Looks like compiling with -Wextra is a good thing to do. This:

	editable.c: In function 'keys_to_str':
	editable.c:130: warning: comparison of unsigned expression < 0 is always false

was triggered by a real bug (obtaining length of a buffer passed as
"char *" argument with sizeof).

In fact, even two bugs: comparison of unsigned expression < 0 is stupid
indeed (and that happens because of C logic we all love).

Signed-off-by: Gregory Petrosyan <gregory.petrosyan@gmail.com>