Add 'track-preferred' and 'album-preferred' for replaygain option.
When replaygain is set to 'track-preferred', album gain will be used
instead if track gain is unavailable. In other words, track gain will be
preferred over album one if available.
Similarly when replaygain is set to 'album-preferred', track gain will be used
instead if album gain is unavailable.
Just in case you want to know, 'disabled, 'track' and 'album' were
only available for replaygain. When replaygain is set to 'track'(or 'album'),
only track gain(or album gain) will be used. If track gain(or album gain) is
unavailable, nothing will be gained.