speedup: use rbtree for sorted and shuffle list
Since the only operations are sorted insert, sorted find (for removal)
and prev/next, a binary tree is much better suited than a flat list.

The shuffle list can completely be replaced by a rbtree (since the
random key is unique), but the sorted list can contain duplicate tracks.
In this case the tree contains the first of the the duplicate tracks.

The Linux 2.6.34 implementation of rbtree is used.

speedup: 98% cpu time for initial indexing
(~30k files, 85% mp3, 14% ogg, 1% other)