pulse: add mixer.pulse.restore_volume option
This option lets cmus restore the last used volume. Because of the
design of pulseaudio it is not possible to change the volume before
playback (or when playback has stopped) anymore.

It is on by default, because setting volume to 100% is very annoying (if
not dangerous to ears) when flat-volumes=yes.

If it is disabled, the old behaviour (before this commit) is activated.

See e.g. Debian bug report 597160:
http://bugs.debian.org/cgi-bin/bugreport.cgi?bug=597160

[gp: document the option + remove the duplication of is_freeform_true()]

Signed-off-by: Gregory Petrosyan <gregory.petrosyan@gmail.com>