id3: fix UTF-16 handling
Cmus can't display tags in e.g. [1] properly (it assumes that since tags are
not in UTF-16BE, they are in UTF-16LE -- they are not, they are in UTF-16 with
BOM).

ID3v2.3 and ID3v2.4 specify that UCS-2 and UTF-16 strings should start with
BOM [2][3]. Unicode standard states that without BOM, you should assume that
string is BE [4].

[1] http://www.lastfm.ru/music/Dahlia%27s+Tear/_/2nd+Sky+Between+Lights%2C+Unreleased+Colours+And+Earth
[2] http://www.id3.org/id3v2.3.0#head-1a37d4a15deafc294208ccfde950f77e47000bca
[3] http://www.id3.org/id3v2.4.0-structure#line-362
[4] http://www.unicode.org/faq/basic_q.html#11

Signed-off-by: Gregory Petrosyan <gregory.petrosyan@gmail.com>