allow overriding charset with CMUS_CHARSET environment variable
On small systems like OpenWrt nl_langinfo(CODESET) always returns
"ASCII". This way the user can override this choice if the terminal
emulator supports more character sets.

This is not a normal option for simplicity: The charset is needed right
at the start of cmus, and changing during runtime is not supported (or
needed?) right now.