Initial CUE sheets support
When adding a file, check for the CUE sheet with the same name. If it exists,
add entries from it instead of the original file.

Of course, this makes the assumption that CUE sheet and track file have the
same name, but this almost always tends to be the case.

Treating CUE sheets as a playlist would have the drawback of adding both the
whole track and individual entries from it in case of recursive directory
import, which is undesireable.

Entries are added as URLs with the following syntax:
cue:///path/to/track.cue/track_n

Signed-off-by: Gregory Petrosyan <gregory.petrosyan@gmail.com>