editable: better datastructure
The rbtree of editable.c now only contains unique tracks, while the
linked list contains all, e.g. for a sorted playlist:

     ________________D__________
    /                .          \
  _B_______          .       ____F_
 / .       \         .      /    . \
A  .        C        .     E     .  G
.  .        .        .     .     .  .
A--B--B--B--C--C--C--D--D--E--E--F--G

For the play queue:

A
.
A--A--A--A--A--A--A

For the library:

     ____D____
    /    .    \
  _B_    .    _F_
 / . \   .   / . \
A  .  C  .  E  .  G
.  .  .  .  .  .  .
A--B--C--D--E--F--G

This should reduce the segfaults due to tree/list degenerations in the
long run. It also simplifies play_queue.c a lot!