add additional format options
This patch adds some additional format options. However, right now they
are not too useful for two reasons:
* Lack of conditionals in format string
  (e.g. we maybe only want to show the composer if one is present)
* Lack of proper handling of multiple values for a tag
  (e.g. we want to join multiple composers, like "John Lennon / Paul
  McCartney")
I plan to address both issues in the future, also the ability to use
them for filtering.

The added format options are:

                     ID3v2  additional freeform
    -------------------------------------------
    %{arranger}
    %{composer}      TCOM   SONGWRITER
    %{conductor}     TPE3
    %{lyricist}      TEXT   SONGWRITER
    %{performer}
    %{remixer}       TPE4
    %{label}         TPUB
    %{publisher}     TPUB
    %{work}
    %{opus}
    %{partnumber}
    %{part}
    %{subtitle}      TIT3   VERSION

Notice that "songwriter" maps to both composer and lyricist, and
"version" is a synonym for "subtitle". Left to do is the ID3v2 mapping
for arranger and performer, which requires multi-value decoding of TMCL
and TIPL. Also "producer" could be added.

Further reading:
http://musicbrainz.org/doc/Picard_Tag_Mapping
http://age.hobba.nl/audio/mirroredpages/ogg-tagging.html

Suggested-by: Leon Merten Lohse <leon@green-side.de>