make enum and array in options.c independent of order
When the enum `format_id` and the array `str_defaults` are used in the functions `options_add` and `options_load`,  both have to be in the same order for the code to work.
The code breaks silently if this is not the case.

This patch removes this dependence and moves the array (which is currently 1000 lines away) right next to the enum to make the connection between these two structures clear.