fix seeking in ffmpeg
When playing with ffmpeg plugin and user seeks to the beginning of track, cmus
wouldn't immediately start from the beginning, but would continue playing for
a while (about 3~5 seconds), and then suddenly jump to the beginning.

A segment of code is added in ffmpeg_seek() to flush the decoded output and
reset the input buffer. This should make it jump immediately, but the current
duration after jump is still not precise.