removes op_jack_drop code as it does not work correctly.
The channels where getting out of sync when drop was called. The
underlying problem is, that can only keep the channels in sync, when
op_jack_cb is the sole consumer of the ring buffers.

If we set a flag and clear the buffers in op_jack_cb, op_jack_drop
returns immediately and new data is written by op_jack_write while or
before (or after) the buffers where cleared. Result is the same as just
dropping the ring buffers.

Returning -OP_ERROR_UNSUPPORTED works fine, just seeking is not as
responsive (but without gaps :)