fixes #124 clear buffers when drop is called
op_jack_drop sets a flag and waits till buffers are dropped in
op_jack_cb. After that it returns control to cmus.

Also the under-run code is made unnecessary by just pausing for one
period (~0.02 seconds with jack params -p 1024 -r 44100) when there is
not enough data in the buffer. Recovery is faster that way. There are
less audible artifacts and the code is prettier.