Make second getch after ESC non-blocking
ncurses already sleeps in the first getch call if it returns an ESC.
Sleeping in the second call is unnecessary because the character (if
any) is already available.

The user can set how long ncurses sleeps by setting the ESCDELAY
environment variable to the time in milliseconds. For example

$ ESCDELAY=25 cmus

will make pressing ESC in command mode leave it almost instantaneously.

Also remove an unnecessary cbreak call because cbreak and halfdelay are
mutually exclusive.