Correct parsing for cue files with byte order mark (#774)
* Correct parsing for cue files with byte order mark

Cue files which have a byte order mark are improperly parsed. Currently,
if there is a byte order mark present and a valid cue token on the first
line of the file, the token is not recognized and the first line of cue
data is missed. For my use case, the first line would the the album
title, and therefore any cue files in my library would be displayed with
`<No Title>` as the album title.

This commit looks for the presence of a UTF-8 byte order mark and
advances the file buffer beyond it. In UTF-8, the byte order mark isn't
recommended and really serves no purpose other than to signify the
beginning of a UTF-8 stream.