Fixed issue with `xstrndup`
Previously, `ape` and a few other modules refused to link, citing a
missing `xstrndup`. This seemed to be due to some issue with
`HAVE_STRNDUP` checks not quite working. Here's a fix. The only issue is
that it may slow down `cmus` a bit since the inline definition (which
was in `xmalloc.h`) has been removed.