Indicate elided text with "…" & related formatting tweaks (#1051)
* Allow width specifier in nested format conditional

... with long form key.

Fixes #637.

* Fix padding calculation when seeing wide character in ascii environment

The wide character was being counted as 2 columns, but is always printed
as "?".

* Support a smaller window size before giving up on formatting

* Change u_copy_chars / u_skip_chars to return remaining width

Also make wide character overskip optional in u_skip_chars, we'll need
this later. (The comment makes sense now, it was documenting code that
hadn't been written yet!)

* Simplify pad/trim code for right-aligned field

* Lead out clipped fields with "..." mark

* Replace format_str implementation and respect format_clipped_text

* Add clipping to remaining window titles

* Change default clipped_text to single character "…" with fallback

* Add format gadget to prioritize from middle of string

* Set default format_current to prioritize the track name

... when the window is too small to display the full information.

* Add max-width formatter for string keys

This gives additional flexibility that might be desirable for
"format_current", e.g. to limit the artist and album to at most 20
characters each:
format_current= %.20a / %.20l%! / %n - %t%= %y

Other types are out of scope, and the intent is not to have full printf
compatibility. We are already incompatible in various ways.

Closes #1004.