Fix segfault when reading mp4 files (#1076)
Fix a problem where any mp4/mpeg-4 file will cause cmus to segfault when read.
This was because of a recent change in `libmp4v2` that changed the MP4Read function from an overload with a filename/atomic callback to a C++ default argument with nullptr, which then caused a segfault when called without the second argument.