Handle colors in active windows correctly (#1120)
Originally, there was the editable_active variable in ui_curses.c to
indicate whether the window is active. However, this variable was
removed in #1051, presumably by mistake. It also leads to the active
variable being uninitialized in print_editable, causing #1093.

The uninitialized variable issues was fixed in #1094, but unfortunately
in a wrong way: it simply assumes all windows are inactive, leading to
incorrect colors in many views.

This commit restores the old logic of editable_active (renamed to
win_active for consistency). Now the colors should be correct.