Add option to sort albums by name in tree view
By default they are still sorted by date, which is the current behavior.

I considered to also add an equivalent of smart_artist_sort, to strip
"The" from album names when sorting. It could be nice to add for
consistency, but I find it's not that useful when the list of albums
under a given artist tends to be quite small anyway.

Closes #909.