Make PART / PARTNUMBER aliases for DISCNUMBER
It seems like that's probably the main intended use, for instance in
id3v2 the disc number would be recorded in TPOS (Part Of Set). The name
"part" is simply more generic as not everything is stored on discs.

As a real world reference, EasyTAG uses PART (and ignores DISCNUMBER)
when reading/writing the "CD" field as an APE tag.

Fixes #1313.