Adding HTTPS support
Most of the changes are contained in http.c and http.h which is to be expected.

Since we have an equivalence between former `int fd;` and ssl` (`read(fd, buf, len) <-> SSL_read(ssl, buf, len)`) we also need to rename many functions that used to take a sockfd as an argument. They can now accept a struct containing both sockfd and ssl struct, along with the pointers to appropriate read/write functions.

We also had to edit parts of input.c because it was heavily relying on sockfd and missing the abstractions we need when connecting over SSL.