Fixes : improved ssl_close() and fixed segfault for local mp3
Removed the wait for close_notify response which is not required by the
specs. This makes ssl_close cleaner.
There was an issue where the fd was closed before the call to
SSL_shutdown().
We somehow expected SSL_CTX_free(ctx) to set ctx to NULL which is not
the case and resulted in a crash.
Only .aac, .mp3 and .ac3 are supported for HLS, the https changes did
not impact other file format so there was a missing initialization of
conn->read and conn->write in the case of local .mp3 files that was
missed.