op/aaudio: Ensure the buffer is at least 80ms
On some devices (e.g., the Pixel 9 with Android 15), the minimum audio
device latency is low enough that the default buffer capacity selected
by AAudio is low enough that cmus can't keep up with producing audio.

This doesn't seem to be an issue on most other devices.

For reference, the default selected capacity and burst size in frames
for 48000 Hz audio was 2886/962 on the Pixel 9, and 6734/3367 on the
Pixel 8.

I chose 80ms as the minimum since it results in AAudio selecting a
capacity and burst size of 4000/2000 on the Pixel 9, and 6734/3367 on
the Pixel 8. Higher values result in slightly lower CPU usage, but
more UI lag. Lower values start running into buffer underruns.