Fix for missing Icecast streaming bitrate display.
If a stream specifies its bitrate in an "icy-br" header, we can use it
to populate the bitrate in the track info. Without this fix, no bitrate
was displayed for a stream, either in the track info or the status bar.

Also changed a print format specifier from %ld to %lld to fix a warning
as the former is too small for the int64_t type.