Do not delete missing files from playlists.
Playlists are important user-supplied data.

Previously, if playlist entries lived on an external hard drive, and
the drive was unplugged, cmus would cause data loss by deleting the
entries from the playlist.

This change fixes that case, ensuring that when users add files to a
playlist, they are never deleted automatically by cmus.